Chess

Description

Bindings for Shiro's chess bot tournament

Documentation

type Bitboard = Word64

A bitboard - an integer where each of the 64 bits represents a square of the board. The bits usually encode the presence of a particular piece. The least significant bit corresponds to a1, the 8th least significant bit corresponds to a8, the most significant bit corresponds to h8.

data Board

A game of chess in a particular state.

data Dir

Directions, used for bitboard manipulation.

Constructors

N

NE

E

SE

S

SW

W

NW

data Piece

Chess piece types

Constructors

Pawn

Bishop

Knight

Rook

Queen

King

data PlayerColor

Player colors

Constructors

White

Black

data State

Game state.

Constructors

Normal
The game is ongoing.

Stalemate
The game has ended in a stalemate.

Checkmate
The game has ended in a checkmate.

data Move

A chess move, specifying the positions but not the piece to be moved.

Constructors

Move

Fields

from :: Bitboard

The move's start position (a bitboard with a single set bit).

to :: Bitboard

The move's end position (a bitboard with a single set bit).

promotion :: Maybe Piece

If this move causes a pawn promotion, specifies the piece to promote it to.

capture :: Bool

Whether this move is a capture.

castle :: Bool

Whether this move is a castle.

moveDefault :: Move

A default Move with all of the fields unset.

IO actions

For info on how to use IO actions, see Haskell wiki.

getBoard :: IO Board

Get the game's current board. The current board changes after every move.

submitMove :: Move -> IO ()

Submit a move that you are going to play.

getTimeMillis :: IO Int

Get the time that was remaining at the start of the turn in milliseconds.

getOpponentTimeMillis :: IO Int

Get opponent's remaining time in milliseconds.

getElapsedTimeMillis :: IO Int

Get time elapsed since the start of the turn in milliseconds.

Board functions

legalMoves :: Board -> [Move]

List legal moves for a board. Note that it may be either your or your opponent's moves depending on what nextTurnColor returns.

nextTurnColor :: Board -> PlayerColor

Returns the player whose turn it currently is.

isWhiteTurn :: Board -> Bool

Whether it's white's turn.

isBlackTurn :: Board -> Bool

Whether it's black's turn.

skipTurn :: Board -> Board

Skips a player's turn, returning an equivalent board where it's the opposite player's turn.

gameState :: Board -> State

Get a game's current state (whether it's a checkmate, stalemate, or neither).

inCheck :: Board -> Bool

Whether the current player is in check.

inCheckmate :: Board -> Bool

Whether the current player is in checkmate.

inDraw :: Board -> Bool

Whether the current player is in a draw.

canCastleKingside :: PlayerColor -> Board -> Bool

Whether a particular player is allowed to castle kingside at some point in the game.

canCastleQueenside :: PlayerColor -> Board -> Bool

Whether a particular player is allowed to castle queenside at some point in the game.

zobristKey :: Board -> Int

Get a board's Zobrist hash.

pushMove :: Move -> Board -> Board

Make a move on a board, returning a new board with that move made.

popMove :: Board -> Board

Undo a move on a board, returning a board without that move.

bitboard :: PlayerColor -> Piece -> Board -> Bitboard

Get a bitboard for a particular player and piece. The bitboard's bits will indicate the presence of that particular piece of that particular color on each square of the board.

pieceFromIndex :: Int -> Board -> Maybe Piece

Get the piece located at a particular index.

pieceFromBitboard :: Bitboard -> Board -> Maybe Piece

Get the piece located at the position set in the bitboard.

colorFromIndex :: Int -> Board -> Maybe PlayerColor

Get the player color for the piece located at a particular index.

colorFromBitboard :: Bitboard -> Board -> Maybe PlayerColor

Get the player color for the piece located at the position set in the bitboard.

Bitboard functions

indexFromBitboard :: Bitboard -> Int

Convert a bitboard with a single set bit to a cell index.

bitboardFromIndex :: Int -> Bitboard

Convert a cell index to a bitboard with a single set bit.

showBitboard :: Bitboard -> String

Get a printable string for a bitboard (a grid with 8 columns and 8 lines)

bbSlide :: Dir -> Bitboard -> Bitboard

Move all elements of the bitboard in a particular direction.

bbFlood

Arguments

:: Bool
Whether to also include the final nonempty space.

-> Dir
Direction to move in.

-> Bitboard
A bitboard representing a set of empty spaces on the board.

-> Bitboard
The bitboard to shift (with a single set bit).

-> Bitboard
The board with movement squares marked. The original square will not be marked.

Travel in a particular direction, setting all bits until encountering a nonempty space, returning all marked spaces.

bbBlocker

Arguments

:: Dir
Direction to move in.

-> Bitboard
A bitboard representing a set of empty spaces on the board.

-> Bitboard
A bitboard to shift (with a single set bit).

-> Bitboard
The board after a shift encountered a nonempty space.

Travel in a particular direction until encountering a nonempty space.

Bitboard function specializations

Specializations for individual directions, matching the C API.

bbSlideN :: Bitboard -> Bitboard

bbSlideS :: Bitboard -> Bitboard

bbSlideE :: Bitboard -> Bitboard

bbSlideW :: Bitboard -> Bitboard

bbSlideNE :: Bitboard -> Bitboard

bbSlideNW :: Bitboard -> Bitboard

bbSlideSE :: Bitboard -> Bitboard

bbSlideSW :: Bitboard -> Bitboard

bbFloodN :: Bool -> Bitboard -> Bitboard -> Bitboard

bbFloodS :: Bool -> Bitboard -> Bitboard -> Bitboard

bbFloodE :: Bool -> Bitboard -> Bitboard -> Bitboard

bbFloodW :: Bool -> Bitboard -> Bitboard -> Bitboard

bbFloodNE :: Bool -> Bitboard -> Bitboard -> Bitboard

bbFloodNW :: Bool -> Bitboard -> Bitboard -> Bitboard

bbFloodSE :: Bool -> Bitboard -> Bitboard -> Bitboard

bbFloodSW :: Bool -> Bitboard -> Bitboard -> Bitboard

bbBlockerN :: Bitboard -> Bitboard -> Bitboard

bbBlockerS :: Bitboard -> Bitboard -> Bitboard

bbBlockerE :: Bitboard -> Bitboard -> Bitboard

bbBlockerW :: Bitboard -> Bitboard -> Bitboard

bbBlockerNE :: Bitboard -> Bitboard -> Bitboard

bbBlockerNW :: Bitboard -> Bitboard -> Bitboard

bbBlockerSE :: Bitboard -> Bitboard -> Bitboard

bbBlockerSW :: Bitboard -> Bitboard -> Bitboard